#!: Quick refactor to split code up neatly
