+++ librjcore/include/rjcore/rjcore.h
@@ -1,59 +1,39 @@
 /**
 
-    \file rjcore/rjcore.h
-    \brief Recycle JTAG Core file
+    @file rjcore/rjcore.h
+    @brief Recycle JTAG Core file
 
 */
 
 #pragma once
 
+#include <stddef.h>
 #include <stdint.h>
+
+#include "rjcoreevent.h"
+#include "rjcoreplatform.h"
 
 #ifdef __cplusplus
 extern "C"
 {
 #endif
 
-/// @brief The platform implementation
-struct RJCorePlatform
-{
-    uint32_t (*currentMillisecondUptime)();    //!< What the current uptime in millisecond is
-    int (*transmitData)(const char *, size_t); //!< Transmits a buffer of data
-};
-
 /// @brief Recycle JTAG core state machine
 struct RJCoreState
 {
-    char void *privateData;
+    struct RJCorePlatform *platform; //!< Platform implementation
+    void *privateData;               //!< User provided private data
 };
 
 /// @brief Initialises the core state machine to start
 /// @param state Pointer to the state to initialise
+/// @param platform Pointer to the platform callbacks
 /// @param privateData Pointer to optional provide data for this implementation
-void RJCoreInit(struct RJCoreState *state, void *privateData);
+void RJCoreInit(struct RJCoreState *state, struct RJCorePlatform *platform, void *privateData);
 
-enum RJCoreEventType
-{
-    RJCoreEventType_ReceivedData, //!< Data has been received over uart
-    RJCoreEventType_Timer,        //!< A timer has been received
-};
-
-struct RJCoreEventReceivedData
-{
-    const char *data;
-    size_t length;
-};
-
-/// @brief An event that has just occurred
-struct RJCoreEvent
-{
-    RJCoreEventType type;
-    union {
-        RJCoreEventReceivedData receivedData;
-    };
-};
-
-void RJCoreNotifyEvent(
+/// @brief Notify the core that an event has occurred
+/// @param event The event that occurred
+void RJCoreNotifyEvent(const struct RJCoreEvent *event);
 
 #ifdef __cplusplus
 }
