+++ platform/virtualdev/runcore.h
@@ -1,0 +1,8 @@
+/*!
+    @file runcore.h
+    @brief Runs the RJCore
+*/
+
+#pragma once
+
+void startRjCore(int fd);
